Refactor DOM shim to external JS and add timeout to Otto
Moved the simulated browser environment (DOM shim) from an inline string in Go to an external setup.js file, embedded via go:embed. Added a script execution timeout and interrupt mechanism to OttoEngine for improved security and stability.
